pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "debgpt.__version__"}

[tool.setuptools.packages.find]
include = ["debgpt*"]
exclude = ["doc*", "tests*"]

[project]
dynamic = ["version"]
name = 'debgpt'
description = "DebGPT -- General Purpose Terminal LLM Tool with Some Debian-Specific Design"
requires-python = ">= 3.10"
dependencies = [
    'rich',
    'prompt_toolkit',
    'openai',
    'requests',
    'bs4',
    'tomli',
    'tenacity',
    'pygments',
    'urwid',
    'lz4',
    'numpy',
]
authors = [
  {name = "Mo Zhou", email = "lumin@debian.org"},
]
readme = "README.md"
license = {text = "LGPL-3.0+"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://salsa.debian.org/deeplearning-team/debgpt"
Issues = "https://salsa.debian.org/deeplearning-team/debgpt/-/issues"

[project.scripts]
debgpt = "debgpt.cli:main"
debgpt-vector-service = "debgpt.vector_service.__main__:main"

[project.optional-dependencies]
dev = [
    "pyflakes",
    "pylint",
    "pypdf",
    "pytest-cov",
    "pytest-timeout",
    "pytest-xdist",
    "pytest>=7.0",
    'anthropic',
    'google-generativeai',
    'pytorch',
    'transformers',
]
vector-service = [
    "fastapi",
    "uvicorn[standard]",
    "qdrant-client",
    "sentence-transformers",
    "sqlalchemy",
    "aiosqlite",
    "python-dotenv",
    "huggingface-hub",
    "openai",
    "google-cloud-aiplatform",
    "requests",
    "numpy",
]
